+++ python/ray/serve/tests/test_http_cancellation.py
@@ -8,12 +8,40 @@ from starlette.requests import Request
 
 import ray
 from ray import serve
-from ray._common.test_utils import Collector, SignalActor, wait_for_condition
+from ray._common.test_utils import SignalActor, wait_for_condition
 from ray.serve._private.test_utils import (
     get_application_url,
     send_signal_on_cancellation,
 )
 from ray.serve.exceptions import RequestCancelledError
+
+
+@ray.remote
+class Collector:
+    def __init__(self):
+        self.items = []
+
+    def add(self, item):
+        self.items.append(item)
+
+    def get(self):
+        return self.items
+
+
+def test_collector_class():
+    collector = Collector.remote()
+
+    random_items = ["this", "is", 1, "demo", "string"]
+
+    for item in random_items:
+        collector.add.remote(item)
+
+    result = ray.get(collector.get.remote())
+
+    assert len(result) == len(random_items)
+
+    for i in range(0, len(result)):
+        assert result[i] == random_items[i]
 
 
 @pytest.mark.parametrize("use_fastapi", [False, True])
